Mudancas feitas 29/04

+++ src/main/java/com/gabrielbl/healthaplication/services/SetorService.java
@@ -2,7 +2,7 @@ package com.gabrielbl.healthaplication.services;
 
 import com.gabrielbl.healthaplication.exception.BusinessException;
 import com.gabrielbl.healthaplication.exception.NotFoundException;
-import com.gabrielbl.healthaplication.model.DTOs.RegisterSetorDTO;
+import com.gabrielbl.healthaplication.model.DTOs.RegistrarSetorDTO;
 import com.gabrielbl.healthaplication.model.Empresa;
 import com.gabrielbl.healthaplication.model.Setor;
 import com.gabrielbl.healthaplication.model.Usuario;
@@ -12,6 +12,8 @@ import com.gabrielbl.healthaplication.repository.UsuarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
 
 @Service
 public class SetorService {
@@ -25,7 +27,7 @@ public class SetorService {
     @Autowired
     private UsuarioRepository usuarioRepository;
 
-    public void criarSetor(RegisterSetorDTO data, String nomeRh) {
+    public void criarSetor(RegistrarSetorDTO data, String nomeRh) {
 
 
         Usuario usuario = usuarioRepository.findByLogin(nomeRh);
@@ -48,9 +50,37 @@ public class SetorService {
     }
 
 
-    public void deletarSetor(RegisterSetorDTO data) {
+    public void deletarSetor(RegistrarSetorDTO data) {
+
         Setor setor = setorRepository.findByNomeAndEmpresaCnpj(data.setor(), data.cnpj());
+
+        if(setor == null) throw new NotFoundException("Setor nao encontrado");
         setorRepository.delete(setor);
 
+
+    }
+
+
+    public void atualizarSetor(RegistrarSetorDTO data, String nomeRh) {
+
+        Usuario usuario = usuarioRepository.findByLogin(nomeRh);
+        if(usuario == null) throw new NotFoundException("Usuario nao encontrado");
+
+        Empresa empresa = usuario.getEmpresa();
+
+        Setor setor = setorRepository.findByNomeAndEmpresaCnpj(empresa.getCnpj(), data.setor());
+
+        setorRepository.save(setor);
+    }
+
+    public List<Setor> getAllSetores() {
+          return setorRepository.findAll();
+    }
+
+    public List<Setor> getAllEmpresaSetores(String cnpj) {
+
+        Empresa empresa = empresaRepository.findByCnpj(cnpj);
+
+        return empresa.getSetor();
     }
 }
